+++ hedera-node/hedera-app/src/main/java/com/hedera/node/app/ServicesMain.java
@@ -56,12 +56,12 @@ import com.swirlds.metrics.api.Metrics;
 import com.swirlds.platform.CommandLineArgs;
 import com.swirlds.platform.builder.PlatformBuilder;
 import com.swirlds.platform.builder.PlatformBuilder.PersistenceScope;
+import com.swirlds.platform.config.ConfigurationSetupUtils;
 import com.swirlds.platform.config.legacy.ConfigurationException;
 import com.swirlds.platform.state.ConsensusStateEventHandler;
 import com.swirlds.platform.state.signed.HashedReservedSignedState;
 import com.swirlds.platform.system.InitTrigger;
 import com.swirlds.platform.system.Platform;
-import com.swirlds.platform.util.BootstrapUtils;
 import com.swirlds.state.State;
 import com.swirlds.state.StateLifecycleManager;
 import com.swirlds.state.merkle.VirtualMapState;
@@ -76,6 +76,7 @@ import org.hiero.base.constructable.ConstructableRegistry;
 import org.hiero.base.constructable.RuntimeConstructable;
 import org.hiero.base.file.FileSystemManager;
 import org.hiero.consensus.config.PathsConfig;
+import org.hiero.consensus.constructable.ConstructableRegistration;
 import org.hiero.consensus.io.RecycleBinImpl;
 import org.hiero.consensus.model.node.NodeId;
 import org.hiero.consensus.roster.ReadableRosterStore;
@@ -140,7 +141,7 @@ public class ServicesMain {
     public static void main(final String... args) throws Exception {
         // --- Configure platform infrastructure and derive node id from the command line and environment ---
         initLogging();
-        BootstrapUtils.setupConstructableRegistry();
+        ConstructableRegistration.setupConstructableRegistry();
         final var commandLineArgs = CommandLineArgs.parse(args);
         if (commandLineArgs.localNodesToStart().size() > 1) {
             logger.error(
@@ -377,7 +378,7 @@ public class ServicesMain {
                 .withSource(SystemEnvironmentConfigSource.getInstance())
                 .withSource(SystemPropertiesConfigSource.getInstance());
 
-        rethrowIO(() -> BootstrapUtils.setupConfigBuilder(
+        rethrowIO(() -> ConfigurationSetupUtils.setupConfigBuilder(
                 configurationBuilder,
                 getAbsolutePath(DEFAULT_SETTINGS_FILE_NAME),
                 getAbsolutePath(DEFAULT_OVERRIDES_YAML_FILE_NAME)));
