+++ platform-sdk/consensus-utility/src/main/java/org/hiero/consensus/constructable/ConstructableRegistration.java
@@ -18,6 +18,17 @@ public final class ConstructableRegistration {
     private ConstructableRegistration() {}
 
     /**
+     * Add all classes to the constructable registry.
+     */
+    public static void setupConstructableRegistry() {
+        try {
+            registerAllConstructables();
+        } catch (final ConstructableRegistryException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    /**
      * Registers {@link Hash}, {@link SerializablePublicKey}, {@link CesEvent}, and {@link NodeId}.
      */
     public static void registerCoreConstructables() throws ConstructableRegistryException {
