+++ platform-sdk/swirlds-cli/src/main/java/org/hiero/consensus/pcli/GenesisPlatformStateCommand.java
@@ -10,7 +10,6 @@ import com.swirlds.common.context.PlatformContext;
 import com.swirlds.config.api.Configuration;
 import com.swirlds.config.api.ConfigurationBuilder;
 import com.swirlds.platform.config.DefaultConfiguration;
-import com.swirlds.platform.util.BootstrapUtils;
 import com.swirlds.state.State;
 import com.swirlds.state.StateLifecycleManager;
 import com.swirlds.state.merkle.VirtualMapStateLifecycleManager;
@@ -19,6 +18,7 @@ import com.swirlds.state.spi.WritableStates;
 import java.io.IOException;
 import java.nio.file.Path;
 import java.util.concurrent.ExecutionException;
+import org.hiero.consensus.constructable.ConstructableRegistration;
 import org.hiero.consensus.model.hashgraph.GenesisSnapshotFactory;
 import org.hiero.consensus.platformstate.PlatformStateAccessor;
 import org.hiero.consensus.roster.RosterStateId;
@@ -56,7 +56,7 @@ public class GenesisPlatformStateCommand extends AbstractCommand {
     @Override
     public Integer call() throws IOException, ExecutionException, InterruptedException, ParseException {
         final Configuration configuration = DefaultConfiguration.buildBasicConfiguration(ConfigurationBuilder.create());
-        BootstrapUtils.setupConstructableRegistry();
+        ConstructableRegistration.setupConstructableRegistry();
 
         final PlatformContext platformContext = PlatformContext.create(configuration);
         final StateLifecycleManager stateLifecycleManager = new VirtualMapStateLifecycleManager(
