+++ platform-sdk/swirlds-cli/src/main/java/org/hiero/consensus/pcli/recovery/EventRecoveryWorkflow.java
@@ -2,7 +2,7 @@
 package org.hiero.consensus.pcli.recovery;
 
 import static com.swirlds.logging.legacy.LogMarker.STARTUP;
-import static com.swirlds.platform.util.BootstrapUtils.setupConstructableRegistry;
+import static org.hiero.consensus.constructable.ConstructableRegistration.setupConstructableRegistry;
 import static org.hiero.consensus.model.PbjConverters.toPbjTimestamp;
 import static org.hiero.consensus.platformstate.PlatformStateUtils.bulkUpdateOf;
 import static org.hiero.consensus.platformstate.PlatformStateUtils.consensusTimestampOf;
@@ -25,7 +25,6 @@ import com.swirlds.platform.system.InitTrigger;
 import com.swirlds.platform.system.SwirldMain;
 import com.swirlds.platform.system.state.notifications.NewRecoveredStateListener;
 import com.swirlds.platform.system.state.notifications.NewRecoveredStateNotification;
-import com.swirlds.platform.util.HederaUtils;
 import com.swirlds.state.StateLifecycleManager;
 import com.swirlds.state.merkle.VirtualMapState;
 import com.swirlds.state.merkle.VirtualMapStateLifecycleManager;
@@ -62,6 +61,7 @@ import org.hiero.consensus.pces.impl.common.PcesFile;
 import org.hiero.consensus.pces.impl.common.PcesMutableFile;
 import org.hiero.consensus.pcli.recovery.internal.EventStreamRoundIterator;
 import org.hiero.consensus.pcli.recovery.internal.StreamedRound;
+import org.hiero.consensus.pcli.utility.HederaUtils;
 import org.hiero.consensus.round.RoundCalculationUtils;
 import org.hiero.consensus.state.SignedStateFileReader;
 import org.hiero.consensus.state.SignedStateFileWriter;
