+++ platform-sdk/swirlds-cli/src/main/java/org/hiero/consensus/pcli/utility/HederaUtils.java
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
-package com.swirlds.platform.util;
+package org.hiero.consensus.pcli.utility;
 
 import com.swirlds.base.time.Time;
 import com.swirlds.common.context.PlatformContext;
